backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ollamastudio-backend"
version = "0.1.0"
description = "OllamaStudio - Backend FastAPI pour interface web Claude Code compatible"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.2",
    "aiofiles>=24.1.0",
    "chardet>=5.2.0",
    "tiktoken>=0.7.0",
    "python-magic>=0.4.27",
    "watchfiles>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
